pro/requirements.txt
altair==5.2.0
anyio==3.6.2
APScheduler==3.6.3
asgiref==3.7.2
attrs==23.2.0
blinker==1.7.0
cachetools==4.2.2
certifi==2022.12.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.0.1
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
cryptography==39.0.0
distlib==0.3.6
dj-database-url==1.2.0
Django==4.2.9
django-braces==1.15.0
django-contrib-comments==2.2.0
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-model-reviews==1.4.0
django-mptt==0.14.0
django-phonenumber-field==7.0.2
django-private-storage==3.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
easyocr==1.7.1
filelock==3.9.0
Flask==3.0.0
flatbuffers==23.5.26
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.41
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
humanfriendly==10.0
idna==3.4
imageio==2.33.1
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpickle==3.0.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
lazy_loader==0.3
llvmlite==0.41.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.3.0
mysqlclient==2.2.1
ndg-httpsclient==0.5.1
networkx==3.2.1
ninja==1.11.1.1
numba==0.58.1
numpy==1.26.3
onnxruntime==1.16.3
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==2.2.0
phonenumberslite==8.13.5
pillow==10.2.0
platformdirs==2.6.2
pooch==1.8.0
protobuf==4.25.1
psycopg2==2.9.5
psycopg2-binary==2.9.5
pusher==3.3.1
pyarrow==15.0.0
pyasn1==0.4.8
pyclipper==1.3.0.post5
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.17.2
PyMatting==1.1.12
PyNaCl==1.5.0
pyOpenSSL==23.0.0
pyreadline3==3.4.1
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python-telegram-bot==13.7
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.32.1
rembg==2.0.53
reportlab==4.0.9
requests==2.28.2
rfc3986==1.5.0
rich==13.7.0
rpds-py==0.16.2
scikit-image==0.22.0
scipy==1.11.4
shapely==2.0.2
six==1.16.0
small-small-hr==0.4.2
smmap==5.0.1
sniffio==1.3.0
sorl-thumbnail==12.9.0
sqlparse==0.4.4
streamlit==1.30.0
streamlit-option-menu==0.3.12
sympy==1.12
tenacity==8.2.3
tifffile==2023.12.9
toml==0.10.2
toolz==0.12.1
torch==2.1.2
torchvision==0.16.2
tornado==6.2
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==4.2
urllib3==1.26.14
validators==0.22.0
virtualenv==20.17.1
voluptuous==0.13.1
watchdog==3.0.0
Werkzeug==3.0.1
whitenoise==6.3.0
zipp==3.17.0
